# backendPI4/api/openapi.yaml
openapi: 3.0.0
info:
  title: API para a aplicação ViriatosScouting
  description:
    "Este documento apresenta toda a documentação de uma API que permite\
    \ gerir um sistema que faz gestão de uma equipa de scouting. A API inclui funcionalidades\
    \ que permitem o utilizador gerir jogadores, clubes, escalões, equipas e utilizadores.\
    \ A mesma tambem apresenta um local para criar tarefas e relatórios."
  version: "1.0"
servers:
  - url: https://localhost:8080
    description: Servidor local utilizado para testar a API
  - url: https://backendpi4-bbzf.onrender.com
    description: Servidor online utilizado para testar a API
...
paths:
...
  /clubes:
    get:
      tags:
        - Clubes
      summary: Lista todos os clubes
      description:
        "Retorna uma lista de todos os clubes, mas somente administradores\
        \ podem visualizar esta informação. Para obter os clubes, o utilizador deve\
        \ estar autenticado com um token válido."
      operationId: listaClubes
      parameters:
        - name: limit
          in: query
          description: Número máximo de clubes a serem retornadas por página.
          required: false
          style: form
          explode: true
          schema:
            maximum: 999999
            minimum: 1
            type: integer
            format: int64
            example: 10
            default: 10
        - name: offset
          in: query
          description: Número de clubes a ignorar para paginação.
          required: false
          style: form
          explode: true
          schema:
            maximum: 300
            minimum: 0
            type: integer
            format: int64
            example: 1
            default: 0
        - name: nome
          in: query
          description: Nome do clube a ir buscar.
          required: false
          style: form
          explode: true
          schema:
            maximum: 30
            minimum: 0
            type: string
            example: "Rocas do Vouga"
      responses:
        "200":
          description: Lista de clubes com metadados de paginação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListarClubes200"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhuma clube encontrado para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorClube"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Clubes
    post:
      tags:
        - Clubes
      summary: Cria um novo clube
      description:
        Permite criar um novo clube. Apenas administradores podem criar
        clubes. O utilizador deve fornecer os dados do clube no body.
      operationId: criarClube
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClubeCriar"
        required: true
      responses:
        "201":
          description: Clube criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubeCriar"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "409":
          description: Conflito - Se o nome da equipa já estiverem
            em uso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Clubes
  /clubes/{id}:
    put:
      tags:
        - Clubes
      summary: Edita um clube
      description:
        Permite editar um clube existente. O utilizador pode editar clubes
        se o mesmo for administrador. É necessário um token válido para autenticação.
      operationId: editarClube
      parameters:
        - name: id
          in: path
          description: ID do clube a ser atualizado
          required: true
          style: simple
          explode: false
          schema:
            maxLength: 100
            pattern: "^[0-9]+$"
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClubeCriar"
        required: true
      responses:
        "200":
          description: Clube atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubeCriar"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhuma clube encontrado para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorClube"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Clubes
    delete:
      tags:
        - Clubes
      summary: Apaga um clube
      description:
        Permite apagar um clube existente. Apenas o administrador pode
        apagá-lo. O utilizador deve fornecer o ID do clube a ser apagado.
      operationId: apagarClube
      parameters:
        - name: id
          in: path
          description: ID do clube a ser apagado
          required: true
          style: simple
          explode: false
          schema:
            maxLength: 100
            pattern: "^[0-9]+$"
            type: string
      responses:
        "204":
          description: Clube apagado com sucesso
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhum clube encontrado para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorClube"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Clubes
...
  /register:
    post:
      tags:
        - Autenticação
      summary: Regista um novo utilizador
      description:
        "Este endpoint permite que um novo utilizador se registe, fornecendo\
        \ o nome de utilizador, email e senha. Após o registo bem-sucedido, um token\
        \ JWT será retornado para autenticação em futuras requisições."
      operationId: registerNormal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Register"
        required: true
      responses:
        "201":
          description: Registo bem-sucedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registar_201"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "409":
          description:
            Conflito - Se o nome de utilizador ou o email já estiverem
            em uso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "429":
          description:
            Muitas requisições - O cliente enviou muitas requisições em
            um curto espaço de tempo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
      x-swagger-router-controller: Autenticao
  /login:
    post:
      tags:
        - Autenticação
      summary: Autentica um utilizador
      description:
        "Este endpoint permite que um utilizador se autentique, fornecendo\
        \ o nome de utilizador e a senha. Se as credenciais forem válidas, será retornado\
        \ um token JWT que pode ser utilizado para autenticação nos próximos endpoints."
      operationId: loginNormal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Login"
        required: true
      responses:
        "200":
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Login_200"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "429":
          description:
            Muitas requisições - O cliente enviou muitas requisições em
            um curto espaço de tempo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
      x-swagger-router-controller: Autenticao
...
components:
  schemas:
...
    Clube:
      required:
        - id
        - nome
      type: object
      properties:
        id:
          maximum: 999999
          minimum: 1
          type: integer
          description: Identificador único do clube
          format: int64
          example: 1
        nome:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nome do clube
          example: Rocas do Vouga
        icon:
          maxLength: 512
          pattern:
            "^(https?|ftp):\\/\\/[\\w\\-]+(\\.[\\w\\-]+)+(\\/[\\w\\-\\.~:@\\\
            /?#\\[\\]@!\\$&'\\(\\)\\*\\+,;=%]*)?$"
          type: string
          description: URL ou caminho para o ícone/imagem do atleta
          example: https://example.com/images/sporting_logo.png
      additionalProperties: false
      example:
        icon: https://example.com/images/sporting_logo.png
        nome: Rocas do Vouga
        id: 1
    ClubeCriar:
      required:
        - nome
      type: object
      properties:
        nome:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- 0-9]+$"
          type: string
          description: Nome do clube
          example: Rocas do Vouga
        icon:
          maxLength: 512
          type: string
          description: URL ou caminho para o ícone/imagem do atleta
          example: https://example.com/images/sporting_logo.png
      additionalProperties: false
...
    Register:
      required:
        - email
        - password
        - nome
      type: object
      properties:
        nome:
          maxLength: 20
          minLength: 3
          pattern: "^[a-zA-ZÀ-ÿ0-9_ ]{3,20}$"
          type: string
          description: Nome de utilizador
          example: gilgraca
        email:
          maxLength: 100
          pattern: "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
          type: string
          description: Endereço de email do utilizador
          format: email
          example: gilgraca@example.com
        password:
          maxLength: 50
          minLength: 8
          pattern: "^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{8,}$"
          type: string
          description: Senha para o novo utilizador
          example: senha-segura123
      additionalProperties: false
    Login:
      required:
        - password
      type: object
      properties:
        nome:
          maxLength: 15
          minLength: 3
          pattern: "^[a-zA-ZÀ-ÿ0-9_ ]{3,20}$"
          type: string
          description: Nome de utilizador para autenticação
          example: gilgraca
        password:
          maxLength: 50
          minLength: 8
          pattern: "^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{8,}$"
          type: string
          description: Palavra-passe do utilizador
          example: P3ssw3rd
      additionalProperties: false
    UnexpectedError:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9À-ÿ ]+$"
          type: string
          example: Ocorreu um erro inesperado no servidor.
      additionalProperties: false
    BadRequestError:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9 ]+$"
          type: string
          example: Token de autenticação inválido ou ausente.
      additionalProperties: false
    UnauthorizedError:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9 ]+$"
          type: string
          example: Campos obrigatórios ausentes ou formato inválido.
      additionalProperties: false
...
    NotFoundErrorClube:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9 ]+$"
          type: string
          description: Mensagem explicativa do erro
          example: Nenhum clube encontrado.
      additionalProperties: false
...
    ConflictError:
      required:
        - conflictDetails
        - error
      type: object
      properties:
        error:
          maxLength: 255
          pattern: '^[a-zA-Z0-9 .,:;!?()''"-]+$'
          type: string
          example: Nome de utilizador ou email já estão em uso.
        conflictDetails:
          $ref: "#/components/schemas/ConflictError_conflictDetails"
      additionalProperties: false
...
    TooManyRequestsError:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9 ]+$"
          type: string
          example:
            Muitas requisições. O cliente enviou muitas requisições em um curto
            espaço de tempo.
      additionalProperties: false
    InternalServerError:
      required:
        - error
      type: object
      properties:
        error:
          maxLength: 255
          pattern: '^[a-zA-Z0-9 .,:;!?()''"-]+$'
          type: string
          example: Ocorreu um erro interno no servidor durante o registo.
        details:
          maxLength: 500
          pattern: '^[a-zA-Z0-9 .,:;!?()''"-]+$'
          type: string
          example: Falha ao conectar com o banco de dados.
        timestamp:
          maxLength: 30
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
          type: string
          format: date-time
          example: 2024-11-11T14:30:00Z
      additionalProperties: false
...
    ListarClubes200:
      type: object
      properties:
        totalCount:
          type: integer
          description: Número total de clubes que correspondem aos filtros aplicados.
        totalPages:
          type: integer
          description: "Número total de páginas, baseado no 'limit' fornecido."
        currentPage:
          type: integer
          description: Página atual (calculada com o offset e limit).
        data:
          maxItems: 50
          type: array
          items:
            $ref: "#/components/schemas/Clube"
      additionalProperties: false
...
    Registar_201:
      type: object
      properties:
        username:
          maxLength: 20
          minLength: 3
          pattern: "^[a-zA-Z0-9_]{3,20}$"
          type: string
          description: Nome de utilizador do novo utilizador
          example: gilgraca
        email:
          maxLength: 100
          pattern: "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
          type: string
          description: Email do novo utilizador
          format: email
          example: gilgraca@example.com
        tipoUtilizador:
          type: integer
          description: Tipo de utilizador
          format: int32
          example: 1
      additionalProperties: false
    Login_200:
      type: object
      properties:
        nome:
          maxLength: 15
          pattern: "^[a-zA-Z0-9]+$"
          type: string
          description: Nome de utilizador
          example: gilgraca
        tipoUtilizador:
          type: integer
          description: Tipo de utilizador
          format: int32
          example: 1
        token:
          maxLength: 300
          pattern: "^[A-Za-z0-9-_]+\\.[A-Za-z0-9-_]+\\.[A-Za-z0-9-_]+$"
          type: string
          description: Token JWT para autenticação
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
      additionalProperties: false
...
    ConflictError_conflictDetails:
      type: object
      properties:
        username:
          maxLength: 15
          pattern: "^[a-zA-Z0-9_]{3,15}$"
          type: string
          example: john_doe
        email:
          maxLength: 100
          pattern: "^[\\w\\.-]+@[\\w\\.-]+\\.\\w{2,}$"
          type: string
          example: john.doe@example.com
      additionalProperties: false
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT